requirements.txt
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
deprecation==2.1.0
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b67
uritemplate==4.1.1
urllib3==2.2.1
